# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-tenant Yemait EduCore school management API
servers:
  - url: /api
    description: Base API path
security:
  - ClerkAuth: []
...
paths:
...
  /parents/{parentId}:
    get:
      operationId: getParent
      tags: [parents]
      summary: Get a parent profile
      parameters:
        - $ref: "#/components/parameters/ParentIdPath"
        - $ref: "#/components/parameters/SchoolId"
      responses:
        "200":
          description: Parent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Parent" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateParent
      tags: [parents]
      summary: Update a parent profile
      parameters:
        - $ref: "#/components/parameters/ParentIdPath"
        - $ref: "#/components/parameters/SchoolId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ParentUpdate" }
      responses:
        "200":
          description: Updated parent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Parent" }
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: Search teachers and non-teaching staff
      parameters:
        - $ref: "#/components/parameters/SchoolId"
        - $ref: "#/components/parameters/Search"
        - name: employeeId
          in: query
          schema: { type: string }
        - name: department
          in: query
          schema: { type: string }
        - name: role
          in: query
          schema: { $ref: "#/components/schemas/EmployeeType" }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/EmployeeStatus" }
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Employee" }
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a teacher or staff member
      parameters:
        - $ref: "#/components/parameters/SchoolId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeInput" }
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
  /employees/{employeeId}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee profile
      parameters:
        - $ref: "#/components/parameters/EmployeeIdPath"
        - $ref: "#/components/parameters/SchoolId"
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee profile
      parameters:
        - $ref: "#/components/parameters/EmployeeIdPath"
        - $ref: "#/components/parameters/SchoolId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeUpdate" }
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
  /employees/{employeeId}/status:
    patch:
      operationId: updateEmployeeStatus
      tags: [employees]
      summary: Change employee status
      parameters:
        - $ref: "#/components/parameters/EmployeeIdPath"
        - $ref: "#/components/parameters/SchoolId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeStatusUpdate" }
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: Clerk session token
      description: Clerk session cookie or bearer token validated by the API
    DeviceCredential:
      type: apiKey
      in: header
      name: X-Device-Credential
      description: Revocable device credential; the raw value is write-only and must never be logged or returned after rotation
  responses:
...
    NotFound:
      description: Resource not found or outside authorized visibility
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  parameters:
    SchoolId:
      name: schoolId
      in: query
      required: true
      schema:
        type: integer
        minimum: 1
...
    ParentIdPath:
      name: parentId
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
    EmployeeIdPath:
      name: employeeId
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
    Search:
      name: search
      in: query
      schema:
        type: string
...
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
...
    Parent:
      type: object
      required: [id, schoolId, name, email, phone, childrenCount, activeChildren]
      properties:
        id: { type: integer }
        schoolId: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        address: { type: ["string", "null"] }
        relationshipType: { type: ["string", "null"], enum: [Father, Mother, Guardian, Grandparent, Other, null] }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
        clerkUserId: { type: ["string", "null"] }
        emergencyContactName: { type: ["string", "null"] }
        emergencyContactPhone: { type: ["string", "null"] }
        childrenCount: { type: integer }
        activeChildren: { type: integer }
...
    ParentUpdate:
      type: object
      properties:
        name: { type: string, minLength: 2 }
        email: { type: string, format: email }
        phone: { type: string, minLength: 7 }
        address: { type: string }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
        emergencyContactName: { type: string }
        emergencyContactPhone: { type: string }
    EmployeeStatus:
      type: string
      enum: [ACTIVE, INACTIVE, SUSPENDED, RESIGNED, TERMINATED]
    EmployeeType:
      type: string
      enum: [TEACHER, STAFF]
    Employee:
      type: object
      required: [id, schoolId, employeeId, firstName, lastName, type, status]
      properties:
        id: { type: integer }
        schoolId: { type: integer }
        employeeId: { type: string }
        firstName: { type: string }
        middleName: { type: ["string", "null"] }
        lastName: { type: string }
        type: { $ref: "#/components/schemas/EmployeeType" }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"], format: email }
        address: { type: ["string", "null"] }
        photoUrl: { type: ["string", "null"], format: uri }
        gender: { type: ["string", "null"], enum: [female, male, other, null] }
        status: { $ref: "#/components/schemas/EmployeeStatus" }
        dateEmployed: { type: ["string", "null"], format: date }
        department: { type: ["string", "null"] }
        qualification: { type: ["string", "null"] }
        userId: { type: ["integer", "null"] }
    EmployeeInput:
      type: object
      required: [employeeId, firstName, lastName, type]
      properties:
        employeeId: { type: string, minLength: 1 }
        firstName: { type: string, minLength: 2 }
        middleName: { type: string }
        lastName: { type: string, minLength: 2 }
        type: { $ref: "#/components/schemas/EmployeeType" }
        phone: { type: string }
        email: { type: string, format: email }
        address: { type: string }
        photoUrl: { type: string, format: uri }
        gender: { type: string, enum: [female, male, other] }
        dateEmployed: { type: string, format: date }
        department: { type: string }
        qualification: { type: string }
        userId: { type: integer, minimum: 1 }
    EmployeeUpdate:
      type: object
      properties:
        firstName: { type: string, minLength: 2 }
        middleName: { type: string }
        lastName: { type: string, minLength: 2 }
        phone: { type: string }
        email: { type: string, format: email }
        address: { type: string }
        photoUrl: { type: string, format: uri }
        gender: { type: string, enum: [female, male, other] }
        dateEmployed: { type: string, format: date }
        department: { type: string }
        qualification: { type: string }
        userId: { type: integer, minimum: 1 }
    EmployeeStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/EmployeeStatus" }
...